+++ src/Test/Hspec/FailureReport.hs
@@ -1,16 +1,21 @@
-module Test.Hspec.FailureReport where
+module Test.Hspec.FailureReport (
+  writeFailureReport
+, readFailureReport
+) where
 
 import           System.IO
 import           System.SetEnv
+import qualified Test.QuickCheck as QC
 import           Test.Hspec.Config
-import           Test.Hspec.Util (safeTry, readMaybe, getEnv)
+import           Test.Hspec.Util (Path, safeTry, readMaybe, getEnv)
 
-writeFailureReport :: String -> IO ()
+type Seed = Integer
+
+writeFailureReport :: (Seed, [Path]) -> IO ()
 writeFailureReport x = do
   -- on Windows this can throw an exception when the input is too large, hence
   -- we use `safeTry` here
-  r <- safeTry (setEnv "HSPEC_FAILURES" x)
-  either onError return r
+  safeTry (setEnv "HSPEC_FAILURES" $ show x) >>= either onError return
   where
     onError err = do
       hPutStrLn stderr ("WARNING: Could not write environment variable HSPEC_FAILURES (" ++ show err ++ ")")
@@ -22,5 +27,12 @@ readFailureReport c = do
     Nothing -> do
       hPutStrLn stderr "WARNING: Could not read environment variable HSPEC_FAILURES; `--re-run' is ignored!"
       return c
-    Just xs -> do
-      return $ configAddFilter (`elem` xs) c
+    Just (seed, xs) -> do
+      (return . setSeed seed . configAddFilter (`elem` xs)) c
+
+setSeed :: Seed -> Config -> Config
+setSeed seed c
+  | hasSeed = c
+  | otherwise = configSetSeed seed c
+  where
+    hasSeed = maybe False (const True) (QC.replay $ configQuickCheckArgs c)